Week-1/chatbot/BE/requirements.txt
fastapi
uvicorn
openai
chromadb
python-docx
PyPDF2
python-dotenv